+++ Source/Libraries/CorruptCore/CorruptCore.csproj
@@ -102,6 +102,7 @@
     <Compile Include="ListFilters.cs" />
     <Compile Include="MemoryBanks.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="BlastLayer.cs" />
     <Compile Include="BlastTools.cs" />
     <Compile Include="CorruptCore.cs" />
     <Compile Include="Diff.cs" />
